requirements.txt
streamlit>=1.27.0
pandas>=1.5.3
openai>=1.12.0

# PDF processing
pdfplumber>=0.10.3
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# DOCX processing
python-docx>=0.8.11

# XLSX processing
openpyxl>=3.1.2

# Email processing
python-magic>=0.4.27

# Utility libraries
tqdm>=4.66.1
numpy>=1.24.3

# Optional: For enhanced OCR capabilities
# tesseract-ocr should be installed separately via system package manager
# On Ubuntu: sudo apt-get install tesseract-ocr
# On macOS: brew install tesseract
# On Windows: Download installer from https://github.com/UB-Mannheim/tesseract/wiki